requirements.txt
asgiref==3.5.1
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
brotlipy @ file:///Users/runner/miniforge3/conda-bld/brotlipy_1636012319596/work
bs4==0.0.1
certifi==2021.10.8
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1636046173594/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1644853463426/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1602866480661/work
conda==4.12.0
conda-package-handling @ file:///Users/runner/miniforge3/conda-bld/conda-package-handling_1647355403437/work
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography_1647857630705/work
Django==4.0.4
django-cors-headers==3.12.0
django-filter==21.1
django-rest-framework==0.1.0
django-webpack-loader==1.5.0
djangorestframework==3.13.1
h11==0.13.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1642433548627/work
image==1.5.33
mysqlclient @ file:///Users/runner/miniforge3/conda-bld/mysqlclient_1639024608475/work
optional-django==0.3.0
outcome==1.1.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
Pillow==9.1.0
pycosat @ file:///Users/runner/miniforge3/conda-bld/pycosat_1636020570107/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1643496850550/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
PyQt5-sip==12.9.0
PySocks @ file:///Users/runner/miniforge3/conda-bld/pysocks_1635862736740/work
pytz==2022.1
react==4.3.0
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1641580202195/work
ruamel-yaml-conda @ file:///Users/runner/miniforge3/conda-bld/ruamel_yaml_1636009262461/work
selenium==4.1.5
sip @ file:///Users/runner/miniforge3/conda-bld/sip_1646101468994/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1604308577558/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1646031859244/work
trio==0.20.0
trio-websocket==0.9.2
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1647489083693/work
wsproto==1.1.0
